Replace out-string with .ToString()
I would like to propose to change the way the code blocks get-converted
to files, so instead of using out-string use the .tostring(). The
reasons for this is that because out-string uses the column-width to
decide where to put in extra line-breaks. So if you have a window with
low width and high zoom the generated script-files in the vm can have
some extra line-breaks that will break the process.
Ps this is my first pull ever, hope I doing this “right”. Thanks for
some great scripts.